FilterSets are not started anymore when loaded. The user has to manually select filters and start them once loaded. This prevents the browser UI from being stuck waiting for the indexing result  because the server is busy doing the initial scan. This results in a better user experience.
The DB content/Filters' rules consistency is now ensured when the indexing results are reloaded directly into the db (when loading a filter set). This is achieved by treating new files as modified files in the filter when invoked by the watcher (see "edge case" comment in Filter::checkNewFile, filter.cpp).